fix(smokes): restore full-public contract parity
