+++ loopx/cli_commands/starter_bootstrap_registration.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import argparse
 
 from ..bootstrap_command_pack import (
-    START_GOAL_CAPABILITY_ROUTES,
     START_GOAL_HOST_SURFACES,
 )
 from ..codex_cli_probe import DEFAULT_CODEX_BIN
@@ -11,17 +10,7 @@ from ..project_prompt import (
     DEFAULT_HANDOFF_ADAPTER_KIND,
     DEFAULT_HANDOFF_ADAPTER_STATUS,
 )
-
-
-def _add_capability_route_argument(parser: argparse.ArgumentParser) -> None:
-    parser.add_argument(
-        "--capability-route",
-        choices=START_GOAL_CAPABILITY_ROUTES,
-        help=(
-            "Explicit product capability route for this goal start. Goal text never "
-            "selects a capability route."
-        ),
-    )
+from .start_goal import add_capability_route_argument, register_start_goal_command
 
 
 def register_starter_bootstrap_commands(subparsers: argparse._SubParsersAction) -> None:
@@ -102,7 +91,7 @@ def register_starter_bootstrap_commands(subparsers: argparse._SubParsersAction) 
         action="append",
         help="Capability available in this host loop. Repeat for multiple capabilities.",
     )
-    _add_capability_route_argument(bootstrap_command_pack_parser)
+    add_capability_route_argument(bootstrap_command_pack_parser)
     bootstrap_command_pack_parser.add_argument(
         "--goal-text",
         help=(
@@ -118,78 +107,7 @@ def register_starter_bootstrap_commands(subparsers: argparse._SubParsersAction) 
         help="Print only the pasteable /loopx handling message.",
     )
 
-    start_goal_parser = subparsers.add_parser(
-        "start-goal",
-        help="Preview a guided /loopx <goal text> start transaction without mutating state.",
-    )
-    start_goal_parser.add_argument(
-        "--guided",
-        action="store_true",
-        help="Required for now: render the guided dry-run transaction packet.",
-    )
-    start_goal_parser.add_argument("--project", default=".", help="Project directory to inspect.")
-    start_goal_parser.add_argument("--goal-id", help="Goal id. Defaults to <project-name>-goal.")
-    start_goal_parser.add_argument(
-        "--agent-id",
-        help=(
-            "Explicit registered LoopX identity for an ongoing session or exact "
-            "user-requested takeover. When omitted, a bound thread identity is reused "
-            "when available; otherwise new onboarding defaults to fresh registration."
-        ),
-    )
-    start_goal_parser.add_argument(
-        "--thread-id",
-        help=(
-            "Stable opaque host thread id used to reuse the bound agent lane. "
-            "Codex App defaults to the ambient CODEX_THREAD_ID when available."
-        ),
-    )
-    start_goal_parser.add_argument(
-        "--new-peer",
-        action="store_true",
-        help="Explicitly request a fresh agent identity for this host thread.",
-    )
-    start_goal_parser.add_argument(
-        "--cli-bin",
-        default="loopx",
-        help="LoopX CLI binary name embedded in generated commands.",
-    )
-    start_goal_parser.add_argument(
-        "--host-surface",
-        choices=START_GOAL_HOST_SURFACES,
-        help=(
-            "Exact host surface that will own loop activation after todo writeback. "
-            "When omitted, start-goal returns a read-only host selection gate."
-        ),
-    )
-    start_goal_parser.add_argument(
-        "--available-capability",
-        dest="available_capabilities",
-        action="append",
-        help="Capability available in this host loop. Repeat for multiple capabilities.",
-    )
-    _add_capability_route_argument(start_goal_parser)
-    goal_input_group = start_goal_parser.add_mutually_exclusive_group(required=True)
-    goal_input_group.add_argument(
-        "--goal-text",
-        help="Exact goal text to plan before todo writeback.",
-    )
-    goal_input_group.add_argument(
-        "--slash-command-arguments",
-        help=(
-            "Complete visible /loopx arguments. The CLI consumes only an optional "
-            "leading --capability-route switch and treats the remainder as goal text. "
-            "Use --slash-command-arguments='<arguments>' when the value begins with --."
-        ),
-    )
-    start_goal_parser.add_argument(
-        "--include-command-pack-detail",
-        action="store_true",
-        help=(
-            "Include the complete nested bootstrap command pack. The default guided "
-            "projection keeps the actionable transaction and advertises this cold path."
-        ),
-    )
+    register_start_goal_command(subparsers)
 
     prompt_parser = subparsers.add_parser(
         "new-project-prompt",
