+++ man/loopx.1
@@ -113,6 +113,9 @@ Create or connect project\-local state.
 \fBloopx new\-project\-prompt\fR
 Generate a copy\-paste project connection prompt for an agent.
 .TP
+\fBloopx bind\-agent\-thread\fR
+Persist one stable host thread binding to an already registered LoopX agent.
+.TP
 \fBloopx codex\-cli\-bootstrap\-message\fR
 Generate the visible Codex CLI TUI setup message.
 .TP
